requirements.txt
requests
flask
PySocks
beautifulsoup4
python-dotenv
